refactor: collapse datetime module into hl7v2
