+++ crates/hl7v2-datetime/Cargo.toml
@@ -13,8 +13,7 @@ keywords = ["hl7", "healthcare", "datetime", "date", "time"]
 categories = ["date-and-time", "parser-implementations"]
 
 [dependencies]
-chrono = { version = "0.4", features = ["serde", "std", "clock"], default-features = false }
-thiserror = { workspace = true }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = ["profile"] }
 
 [dev-dependencies]
 proptest = "1.6"
